Split native runtime test support
